requirements.txt
Pillow>=10.0.0
Jinja2>=3.1.0
ReportLab>=4.0.0
svgwrite>=1.4.3
openpyxl>=3.1.2
